Lib

Documentation

splitByWhiteChars

Arguments

:: String
String that will be splited.

-> [String]
List of words made by splitting.

Splits given string by all white characters.

For example:

>>> splitByWhiteChars "The \n\t whit\ne chars in\t this sentence looks\n strange\t."
["The","","","","whit","e","chars","in","","this","sentence","looks","","strange","."]

filterEmpty

Arguments

:: [String]
List to be filtered.

-> [String]
Filtered list.

Filters empty strings from a list.

For example:

>>> filterEmpty ["The","","","","whit","e","chars","in","","this","sentence","looks","","strange","."]
["The","whit","e","chars","in","this","sentence","looks","strange","."]

splitByOperators

Arguments

:: [String]
List of words that need to splited by operators.

-> [String]
List of words and separated operators.

Splits strings from a given list by cpp operators [eg. +] and returns splited words lists concatenated.

For example:

>>> splitByOperators ["ala+ma-kota", "cout<<should+it+=work+=like==this?"]
["ala","+","ma","-","kota","cout","<<","should","+","it","+=","work","+=","like","==","this?"]

printFormated

Arguments

:: [String]
List of words to be printed into new string.

-> Int
Number of tabs for actual row.

-> Bool
Is first element first in new line?

-> String
Concatenated input strings with right successorss and predecessors attached for them.

Adds to each word from list its successors and predecessors and tabs and concatenates all in one string.

For example:

>>> printFormated [ "void", "main", "(", ")", "{", "return", "0", ";", "}"] 0 True
"\nvoid main( ) {\n\t return 0;\n }\n"